simplify animation / update_modals
