+++ webapp/src/actions.ts
@@ -9,7 +9,8 @@ import {addChannelMember} from 'mattermost-redux/actions/channels';
 import {DispatchFunc, GetStateFunc} from 'mattermost-redux/types/actions';
 import {getCurrentChannelId} from 'mattermost-webapp/packages/mattermost-redux/src/selectors/entities/common';
 
-import {makeModalDefinition as makeUpdateRunModalDefinition} from 'src/components/modals/run_update';
+import {makeModalDefinition as makeUpdateRunNameModalDefinition} from 'src/components/modals/run_update_name';
+import {makeModalDefinition as makeUpdateRunChannelModalDefinition} from 'src/components/modals/run_update_channel';
 import {makeModalDefinition as makePlaybookRunModalDefinition} from 'src/components/modals/run_playbook_modal';
 import {makeModalDefinition as makePlaybookRunNewModalDefinition} from 'src/components/modals/new_run_playbook_modal';
 import {PlaybookRun} from 'src/types/playbook_run';
@@ -106,13 +107,19 @@ export function startPlaybookRun(teamId: string, postId?: string) {
     };
 }
 
-export function openUpdateRunModal(playbookRunId: string, teamId: string, field: 'name' | 'channel_id', onSubmitStart?: () => void, onSubmitEnd?: (successMessage?: string) => void) {
-    return modals.openModal(makeUpdateRunModalDefinition({
+export function openUpdateRunNameModal(playbookRunId: string, teamId: string, onSubmit: (newName: string) => void) {
+    return modals.openModal(makeUpdateRunNameModalDefinition({
         playbookRunId,
         teamId,
-        field,
-        onSubmitStart,
-        onSubmitEnd,
+        onSubmit,
+    }));
+}
+
+export function openUpdateRunChannelModal(playbookRunId: string, teamId: string, onSubmit: (newChannelId: string, newChannelName: string) => void) {
+    return modals.openModal(makeUpdateRunChannelModalDefinition({
+        playbookRunId,
+        teamId,
+        onSubmit,
     }));
 }
